i2c: tegra: Fix i2c unknown interrupt issue
Writes to modules on APB bus may complete out-of-order with respect
to the GIC. This can happen when we clear the latched interrupt status
just prior to leaving the interrupt handle and having the GIC driver
unmask the irq. Since this can post before the i2c device write
behind the APB, we can get into the situation where the interrupt
is still pending in the GIC, and the irq handler gets called right
back. We need to guarantee that the write to i2c completes before
leaving the interrupt handle by reading the status register back.

bug 980763

Change-Id: Ibe7070fa6509b1b4d1f510ac2a2ef88a6ac79a8d
Signed-off-by: Haley Teng <hteng@nvidia.com>
Signed-off-by: Dima Zavin <dima@android.com>